refactor: separating user model
