+++ internal/model/detailsModel.go
@@ -1,0 +1,8 @@
+package model
+
+type CreateNewDetails struct {
+	Departments string `json:"departments"`
+	Directions  string `json:"directions"`
+	JobNumber   string `json:"jobNumber"`
+	Contact     string `json:"contact"`
+}
