+++ internal/model/userModel.go
@@ -1,7 +1,5 @@
 package model
 
-import "fmt"
-
 type UserData struct {
 	Rut        string `json:"rut"`
 	Names      string `json:"names"`
@@ -46,13 +44,6 @@ type CreateNewUser struct {
 	Role      string `json:"role"`
 }
 
-type CreateNewDetails struct {
-	Departments string `json:"departments"`
-	Directions  string `json:"directions"`
-	JobNumber   string `json:"jobNumber"`
-	Contact     string `json:"contact"`
-}
-
 type UpdateUserInput struct {
 	Rut         *string `json:"name"`
 	Names       *string `json:"names"`
@@ -75,21 +66,3 @@ type RegisterUser struct {
 	JobNumber   *string `json:"jobNumber" binding:"required"`
 	Contact     *string `json:"contact" binding:"required"`
 }
-
-type LoginData struct {
-	Email    string `json:"email" binding:"required"`
-	Password string `json:"password" binding:"required"`
-}
-
-type LoginResponse struct {
-	Message    string
-	Token      string
-	Names      string
-	Email      string
-	Role       string
-	Identifier int
-}
-
-var ErrUserNotFound = fmt.Errorf("user not found")
-var ErrToken = fmt.Errorf("couldnt complete login, try again later")
-var ErrIncorrectPassword = fmt.Errorf("Incorrect password, try again")
